segregated vaos
