+++ Bess/include/scene/renderer/gl/vao.h
@@ -126,256 +126,4 @@ namespace Bess::Gl {
             0, 1, 2, // First triangle
         };
     } // namespace
-
-    template <typename VertexType>
-    class InstancedVao {
-      public:
-        InstancedVao(size_t maxInstances) {
-            m_vao = std::make_unique<Vao>();
-            m_templateVbo = std::make_unique<Vbo>(QuadTemplateVertices.size() * sizeof(TemplateVertex), QuadTemplateVertices.data());
-            m_vao->addVertexBuffer(*m_templateVbo, getTemplateLayout());
-
-            m_ibo = std::make_unique<Ibo>(QuadIndices.size(), QuadIndices.data());
-            m_vao->setIndexBuffer(*m_ibo);
-
-            m_instanceVbo = std::make_unique<Vbo>(maxInstances * sizeof(VertexType));
-            m_vao->addVertexBuffer(*m_instanceVbo, getInstanceLayout());
-        }
-
-        InstancedVao(size_t maxInstances, BufferLayout instanceLayout) {
-            m_vao = std::make_unique<Vao>();
-            m_templateVbo = std::make_unique<Vbo>(QuadTemplateVertices.size() * sizeof(TemplateVertex), QuadTemplateVertices.data());
-            m_vao->addVertexBuffer(*m_templateVbo, getTemplateLayout());
-
-            m_ibo = std::make_unique<Ibo>(QuadIndices.size(), QuadIndices.data());
-            m_vao->setIndexBuffer(*m_ibo);
-
-            m_instanceVbo = std::make_unique<Vbo>(maxInstances * sizeof(VertexType));
-            m_vao->addVertexBuffer(*m_instanceVbo, instanceLayout);
-        }
-
-        InstancedVao(size_t maxInstances, size_t templateVerticiesSize,
-                     void *templateVertices, size_t templateIndiciesSize, GLuint *templateIndicies) {
-            m_vao = std::make_unique<Vao>();
-            m_templateVbo = std::make_unique<Vbo>(templateVerticiesSize, templateVertices);
-            m_vao->addVertexBuffer(*m_templateVbo, getTemplateLayout());
-
-            m_ibo = std::make_unique<Ibo>(templateIndiciesSize, templateIndicies);
-            m_vao->setIndexBuffer(*m_ibo);
-
-            m_instanceVbo = std::make_unique<Vbo>(maxInstances * sizeof(VertexType));
-            m_vao->addVertexBuffer(*m_instanceVbo, getInstanceLayout());
-        }
-
-        ~InstancedVao() = default;
-
-        InstancedVao(const InstancedVao &) = delete;
-        InstancedVao &operator=(const InstancedVao &) = delete;
-        InstancedVao(InstancedVao &&) = delete;
-        InstancedVao &operator=(InstancedVao &&) = delete;
-
-        void bind() const {
-            m_vao->bind();
-        }
-
-        void unbind() const {
-            m_vao->unbind();
-        }
-
-        virtual BufferLayout getTemplateLayout() {
-            return BufferLayout{
-                {GL_FLOAT, 2, sizeof(TemplateVertex), offsetof(TemplateVertex, position), false}, // a_LocalPosition
-                {GL_FLOAT, 2, sizeof(TemplateVertex), offsetof(TemplateVertex, texCoord), false}  // a_TexCoord
-            };
-        };
-
-        virtual BufferLayout getInstanceLayout() {
-            return BufferLayout{
-                {GL_FLOAT, 3, sizeof(InstanceVertex), offsetof(InstanceVertex, position), false, true},
-                {GL_FLOAT, 2, sizeof(InstanceVertex), offsetof(InstanceVertex, size), false, true},
-                {GL_FLOAT, 1, sizeof(InstanceVertex), offsetof(InstanceVertex, angle), false, true},
-                {GL_FLOAT, 4, sizeof(InstanceVertex), offsetof(InstanceVertex, color), false, true},
-                {GL_INT, 1, sizeof(InstanceVertex), offsetof(InstanceVertex, id), false, true},
-                {GL_INT, 1, sizeof(InstanceVertex), offsetof(InstanceVertex, texSlotIdx), false, true},
-                {GL_FLOAT, 4, sizeof(InstanceVertex), offsetof(InstanceVertex, texData), false, true},
-            };
-        }
-
-        void updateInstanceData(const std::vector<VertexType> &vertexData) {
-            m_instanceVbo->setData(vertexData.data(), vertexData.size() * sizeof(VertexType));
-        }
-
-        size_t getIndexCount() const {
-            if (m_ibo) {
-                return m_ibo->getCount();
-            }
-            return 0;
-        }
-
-      protected:
-        std::unique_ptr<Vao> m_vao;
-        std::unique_ptr<Vbo> m_templateVbo;
-        std::unique_ptr<Ibo> m_ibo;
-        std::unique_ptr<Vbo> m_instanceVbo;
-    };
-
-    class QuadVao : public InstancedVao<QuadVertex> {
-      public:
-        QuadVao(size_t limit) : InstancedVao(limit, getInstanceLayout()) {}
-        BufferLayout getInstanceLayout() override;
-    };
-
-    class CircleVao : public InstancedVao<CircleVertex> {
-      public:
-        CircleVao(size_t limit) : InstancedVao(limit, getInstanceLayout()) {}
-        BufferLayout getInstanceLayout() override;
-    };
-    
-    template<typename VertexType>
-    class BatchVao {
-      public:
-        BatchVao(size_t maxEntities, size_t verticesPerEntt, size_t indicesPerEntt, bool genIndicies = true, bool isQuad = true) : m_isQuad(isQuad) {
-            m_vao = std::make_unique<Vao>();
-
-            m_vbo = std::make_unique<Vbo>(sizeof(VertexType) * maxEntities * verticesPerEntt);
-            m_vao->addVertexBuffer(*m_vbo, getVertexLayout());
-            
-
-            GLuint *indicies = nullptr;
-            if (genIndicies) {
-                auto info = getIndiciesInfo();
-                indicies = genIndiciesData(maxEntities, info.first, info.second);
-            }
-
-            // passing count here not size, need to fix this
-            m_ibo = std::make_unique<Ibo>(maxEntities * indicesPerEntt, indicies);
-            m_vao->setIndexBuffer(*m_ibo);
-
-            if (indicies != nullptr)
-                delete indicies;
-        }
-
-		~BatchVao() = default;
-        BatchVao(const BatchVao &) = delete;
-        BatchVao &operator=(const BatchVao &) = delete;
-        BatchVao(BatchVao &&) = delete;
-        BatchVao &operator=(BatchVao &&) = delete;
-        
-        virtual BufferLayout getVertexLayout() {
-            return BufferLayout{
-                {GL_FLOAT, 3, sizeof(Vertex), offsetof(Vertex, position), false, false},
-                {GL_FLOAT, 4, sizeof(Vertex), offsetof(Vertex, color), false, false},
-                {GL_FLOAT, 2, sizeof(Vertex), offsetof(Vertex, texCoord), false, false},
-                {GL_INT, 1, sizeof(Vertex), offsetof(Vertex, id), false, false},
-                {GL_INT, 1, sizeof(Vertex), offsetof(Vertex, texSlotIdx), false, false},
-            };
-        }
-        
-        virtual std::pair<int, const std::vector<GLuint>> getIndiciesInfo() {
-            if (m_isQuad) {
-                return {
-                    4,
-                    {0, 1, 2, 2, 3, 0}
-                };
-            } else {
-                return {
-                    3,
-                    {0, 1, 2}
-                };
-            }
-        }
-
-        void bind() const {
-            m_vao->bind();
-        }
-
-        void unbind() const {
-            m_vao->unbind();
-        }
-
-        void setVertices(const std::vector<VertexType> &verticies) {
-            m_vbo->setData(verticies.data(), verticies.size() * sizeof(Vertex));
-        }
-
-        void setVerticesAndIndicies(const std::vector<VertexType>& verticies, const std::vector<GLuint>& indicies){
-            m_vbo->setData(verticies.data(), verticies.size() * sizeof(VertexType));
-            m_ibo->setData(indicies.data(), indicies.size() * sizeof(GLuint));
-        }
-
-        private:
-        GLuint *genIndiciesData(
-            size_t maxEntt, 
-            int increment, 
-            const std::vector<GLuint>& enttIndicies // indicies of single entity
-        ) {
-            size_t len = enttIndicies.size();
-            size_t maxIndicies = len * maxEntt;
-            GLuint *indicies = new GLuint[maxIndicies];
-            for (int i = 0; i < len; i++)
-                indicies[i] = enttIndicies[i];
-
-            for (size_t i = len; i < maxIndicies; i++) {
-                indicies[i] = indicies[i - len] + increment;
-            }
-
-            return indicies;
-        }
-
-      protected:
-        std::unique_ptr<Vao> m_vao;
-        std::unique_ptr<Ibo> m_ibo;
-        std::unique_ptr<Vbo> m_vbo;
-        bool m_isQuad;
-    };
-
-
-    class GridVao {
-      public:
-        GridVao() {
-            m_vao = std::make_unique<Vao>();
-
-            m_vbo = std::make_unique<Vbo>(sizeof(GridVertex) * 4);
-            m_vao->addVertexBuffer(*m_vbo, getVertexLayout());
-
-            GLuint indicies[] = {0, 1, 2, 2, 3, 0};
-            m_ibo = std::make_unique<Ibo>(6, indicies);
-            m_vao->setIndexBuffer(*m_ibo);
-        }
-
-        void bind() const {
-            m_vao->bind();
-        }
-
-        void unbind() const {
-            m_vao->unbind();
-        }
-
-        void setVertices(const std::vector<GridVertex> &verticies) {
-            m_vbo->setData(verticies.data(), verticies.size() * sizeof(Vertex));
-        }
-
-        BufferLayout getVertexLayout() {
-            return BufferLayout{
-                {GL_FLOAT, 3, sizeof(GridVertex), offsetof(GridVertex, position), false, false},
-                {GL_FLOAT, 2, sizeof(GridVertex), offsetof(GridVertex, texCoord), false, false},
-                {GL_INT  , 1, sizeof(GridVertex), offsetof(GridVertex, id), false, false},
-                {GL_FLOAT, 4, sizeof(GridVertex), offsetof(GridVertex, color), false, false},
-                {GL_FLOAT, 1, sizeof(GridVertex), offsetof(GridVertex, ar), false, false},
-            };
-        }
-      protected:
-        std::unique_ptr<Vao> m_vao;
-        std::unique_ptr<Ibo> m_ibo;
-        std::unique_ptr<Vbo> m_vbo;
-        bool m_isQuad;
-    };
-
-    class TriangleVao : public InstancedVao<InstanceVertex> {
-      public:
-        TriangleVao(size_t limit) : InstancedVao(limit, 
-            sizeof(TriangleTemplateVertices), (void*)TriangleTemplateVertices.data(),
-            sizeof(TriangleTemplateIndices), (GLuint*)TriangleTemplateIndices.data()
-        ) {}
-    };
-
 } // namespace Bess::Gl
